requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
Django==3.2.20
django-cors-headers==4.1.0
djangorestframework==3.14.0
idna==3.4
praw==7.7.1
prawcore==2.3.0
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.7.1
update-checker==0.18.0
urllib3==2.0.4
websocket-client==1.6.1
